--- src/phoenix/datasets/errors.py
@@ -1,16 +1,0 @@
-from __future__ import annotations
-
-
-class EvaluatorImplementationError(Exception): ...
-
-
-class EvaluatorHasPositionalOnlyParameter(EvaluatorImplementationError): ...
-
-
-class EvaluatorHasInvalidParameterName(EvaluatorImplementationError): ...
-
-
-class EvaluatorIsMissingVariadicKeywordParameters(EvaluatorImplementationError): ...
-
-
-class EvaluatorHasInvalidReturnAnnotation(EvaluatorImplementationError): ...
