+++ src/phoenix/datasets/evaluators/utils.py
@@ -1,12 +1,16 @@
 import functools
 import inspect
-from typing import Any, Callable, Optional, Union
+from abc import ABC
+from types import MappingProxyType
+from typing import Any, Awaitable, Callable, Mapping, Optional, Union
+
+from typing_extensions import TypeAlias
 
 from phoenix.datasets.types import (
     AnnotatorKind,
     EvaluationResult,
-    Evaluator,
     JSONSerializable,
+    TaskOutput,
 )
 
 
@@ -22,7 +26,7 @@ def _unwrap_json(obj: JSONSerializable) -> JSONSerializable:
     return obj
 
 
-def _validate_signature(sig: inspect.Signature) -> None:
+def validate_signature(sig: inspect.Signature) -> None:
     # Check that the wrapped function has a valid signature for use as an evaluator
     # If it does not, raise an error to exit early before running evaluations
     params = sig.parameters
@@ -30,8 +34,9 @@ def _validate_signature(sig: inspect.Signature) -> None:
     if len(params) == 0:
         raise ValueError("Evaluation function must have at least one parameter.")
     if len(params) > 1:
-        not_found = set(params) - valid_named_params
-        if not_found:
+        for not_found in set(params) - valid_named_params:
+            if params[not_found].kind is inspect.Parameter.VAR_KEYWORD:
+                continue
             raise ValueError(
                 (
                     f"Invalid parameter names in evaluation function: {', '.join(not_found)}. "
@@ -63,7 +68,7 @@ def create_evaluator(
     kind: Union[str, AnnotatorKind] = AnnotatorKind.CODE,
     name: Optional[str] = None,
     scorer: Optional[Callable[[Any], EvaluationResult]] = None,
-) -> Callable[[Callable[..., Any]], Evaluator]:
+) -> Callable[[Callable[..., Any]], "Evaluator"]:
     if scorer is None:
         scorer = _default_eval_scorer
 
@@ -82,7 +87,7 @@ def create_evaluator(
         assert name is not None
 
         wrapped_signature = inspect.signature(func)
-        _validate_signature(wrapped_signature)
+        validate_signature(wrapped_signature)
 
         if inspect.iscoroutinefunction(func):
             return _wrap_coroutine_evaluation_function(name, kind, wrapped_signature, scorer)(func)
@@ -97,8 +102,8 @@ def _wrap_coroutine_evaluation_function(
     annotator_kind: AnnotatorKind,
     sig: inspect.Signature,
     convert_to_score: Callable[[Any], EvaluationResult],
-) -> Callable[[Callable[..., Any]], Evaluator]:
-    def wrapper(func: Callable[..., Any]) -> Evaluator:
+) -> Callable[[Callable[..., Any]], "Evaluator"]:
+    def wrapper(func: Callable[..., Any]) -> "Evaluator":
         class AsyncEvaluator(Evaluator):
             def __init__(self) -> None:
                 self._name = name
@@ -123,8 +128,8 @@ def _wrap_sync_evaluation_function(
     annotator_kind: AnnotatorKind,
     sig: inspect.Signature,
     convert_to_score: Callable[[Any], EvaluationResult],
-) -> Callable[[Callable[..., Any]], Evaluator]:
-    def wrapper(func: Callable[..., Any]) -> Evaluator:
+) -> Callable[[Callable[..., Any]], "Evaluator"]:
+    def wrapper(func: Callable[..., Any]) -> "Evaluator":
         class SyncEvaluator(Evaluator):
             def __init__(self) -> None:
                 self._name = name
@@ -153,3 +158,130 @@ def _default_eval_scorer(result: Any) -> EvaluationResult:
         return result
     else:
         raise ValueError(f"Unsupported evaluation result type: {type(result)}")
+
+
+ExampleOutput: TypeAlias = Mapping[str, JSONSerializable]
+ExampleMetadata: TypeAlias = Mapping[str, JSONSerializable]
+ExampleInput: TypeAlias = Mapping[str, JSONSerializable]
+
+EvaluatorName: TypeAlias = str
+EvaluatorKind: TypeAlias = str
+EvaluatorOutput: TypeAlias = Union[EvaluationResult, bool, int, float, str]
+
+
+class Evaluator(ABC):
+    """
+    A helper super class to guide the implementation of an `Evaluator` object.
+    Subclasses must implement either the `evaluate` or `async_evaluate` method.
+    Implementing both methods is recommended, but not required.
+
+    This Class is intended to be subclassed, and should not be instantiated directly.
+    """
+
+    _kind: AnnotatorKind
+    _name: EvaluatorName
+
+    @functools.cached_property
+    def name(self) -> EvaluatorName:
+        if hasattr(self, "_name"):
+            return self._name
+        return self.__class__.__name__
+
+    @functools.cached_property
+    def kind(self) -> EvaluatorKind:
+        if hasattr(self, "_kind"):
+            return self._kind.value
+        return AnnotatorKind.CODE.value
+
+    def __new__(cls, *args: Any, **kwargs: Any) -> "Evaluator":
+        if cls is Evaluator:
+            raise TypeError(f"{cls.__name__} is an abstract class and should not be instantiated.")
+        return object.__new__(cls)
+
+    def evaluate(
+        self,
+        *,
+        output: Optional[TaskOutput] = None,
+        expected: Optional[ExampleOutput] = None,
+        metadata: ExampleMetadata = MappingProxyType({}),
+        input: ExampleInput = MappingProxyType({}),
+        **kwargs: Any,
+    ) -> EvaluationResult:
+        # For subclassing, one should implement either this sync method or the
+        # async version. Implementing both is recommended but not required.
+        raise NotImplementedError
+
+    async def async_evaluate(
+        self,
+        *,
+        output: Optional[TaskOutput] = None,
+        expected: Optional[ExampleOutput] = None,
+        metadata: ExampleMetadata = MappingProxyType({}),
+        input: ExampleInput = MappingProxyType({}),
+        **kwargs: Any,
+    ) -> EvaluationResult:
+        # For subclassing, one should implement either this async method or the
+        # sync version. Implementing both is recommended but not required.
+        return self.evaluate(
+            output=output,
+            expected=expected,
+            metadata=metadata,
+            input=input,
+            **kwargs,
+        )
+
+    def __init_subclass__(cls, is_abstract: bool = False, **kwargs: Any) -> None:
+        super().__init_subclass__(**kwargs)
+        if is_abstract:
+            return
+        evaluate_fn_signature = inspect.signature(Evaluator.evaluate)
+        for super_cls in inspect.getmro(cls):
+            if super_cls in (LLMEvaluator, Evaluator):
+                break
+            if evaluate := super_cls.__dict__.get(Evaluator.evaluate.__name__):
+                assert callable(evaluate), "`evaluate()` method should be callable"
+                # need to remove the first param, i.e. `self`
+                _validate_sig(functools.partial(evaluate, None), "evaluate")
+                return
+            if async_evaluate := super_cls.__dict__.get(Evaluator.async_evaluate.__name__):
+                assert callable(async_evaluate), "`async_evaluate()` method should be callable"
+                # need to remove the first param, i.e. `self`
+                _validate_sig(functools.partial(async_evaluate, None), "async_evaluate")
+                return
+        raise ValueError(
+            f"Evaluator must implement either "
+            f"`def evaluate{evaluate_fn_signature}` or "
+            f"`async def async_evaluate{evaluate_fn_signature}`"
+        )
+
+
+def _validate_sig(fn: Callable[..., Any], fn_name: str) -> None:
+    sig = inspect.signature(fn)
+    validate_signature(sig)
+    for param in sig.parameters.values():
+        if param.kind is inspect.Parameter.VAR_KEYWORD:
+            return
+    else:
+        raise ValueError(f"`{fn_name}` should allow variadic keyword arguments `**kwargs`")
+
+
+class LLMEvaluator(Evaluator, ABC, is_abstract=True):
+    """
+    A convenience super class for setting `kind` as LLM.
+
+    This Class is intended to be subclassed, and should not be instantiated directly.
+    """
+
+    _kind = AnnotatorKind.LLM
+
+    def __new__(cls, *args: Any, **kwargs: Any) -> "LLMEvaluator":
+        if cls is LLMEvaluator:
+            raise TypeError(f"{cls.__name__} is an abstract class and should not be instantiated.")
+        return object.__new__(cls)
+
+
+ExperimentEvaluator: TypeAlias = Union[
+    Evaluator,
+    Callable[..., EvaluatorOutput],
+    Callable[..., Awaitable[EvaluatorOutput]],
+]
